docker: Mount and clone local swh-jenkins-jobs repo for conf testing
Currently, the only way to test Jenkins configuration changes using
docker is to push changes to a remote feature branch in swh-jenkins-jobs
repository and make jenkins clone that branch to test the changes.

This is not really convenient to quickly test changes so prefer to
mount the local swh-jenkins-jobs repository as a docker volume
and clone that repository from Jenkins filesystem.