templates/swh-pipeline: Fetch master branch only when cloning repo
The jobs associated to this pipeline are only triggered when commiting
on the master branch or when pushing a phabricator diff based on a
commit of the master branch.

So in order to save time, data transfer and avoid possible clone issues,
fetch only the master branch when cloning as all remote branches are
fetched by default.

Related to T4641