templates/stage-cypress-tests: Prevent race condition with support file
Sometimes when multiple cypress processes load the same support file
in parallel, a race condition is encountered as file copy operations
are performed. As a result, the content of the support file is not
loaded and cypress tests will fail.

So use a specific copy of the support file for each cypress runner
to mitigate that issue.