jobs: Add gitlab triggerable jobs configuration for swh packages
For each swh package, add a "gitlab builds" job that will be triggered when
pushing commits or creating merge requests on a swh package repository hosted
on a swh GitLab instance.

Those jobs are based on the use of Jenkins integration on the GitLab side
and on the use of the GitLab plugin on the Jenkins side.