jobs: Prefer to inline utility function instead of using shared lib
Currently we are importing a global shared groovy library stored in the
https://gitlab.softwareheritage.org/swh/infra/ci-cd/swh-jenkins-library
git repository and which only contains a simple function named jobExists.

Importing that shared library messes up with the git-forensics Jenkins
plugin used by the code-coverage-api Jenkins plugin to compute the
coverage on changes from a merge request. Indeed the head commit of
the swh-jenkins-library is used instead of the one from the repository
where a merge request was submitted.
See https://issues.jenkins.io/browse/JENKINS-68675.

As that function is only used in a couple of jobs, prefer to inline it
instead of importing the global shared library in all jobs.