jobs/swh-pipeline: Change executed job when pushing to master
In order to have a code coverage report focused on the changes of a merge
request displayed in the Jenkins UI, prefer to execute the {name}/master
job instead of the {name}/gitlab-builds one.

Indeed the {name}/gitlab-builds is triggered by pushing branches, tags
or merge requests which can confuse the git-forensics Jenkins plugin.
It is used to compute relative results that show what will be changed
if the  branch of the current build will be merged with the master branch.

By ensuring only {name}/master is executed when pushing to the master
branch, finding the reference commit a merge request is based on and its
associated build job is guaranteed.