Add job to build docker images impacted by a new swh release
This opens a new job build-docker-images to allow impacted docker images to (re)build
after a new release.

For now, the job is triggered manually.

In effect, it compulses the list of impacted images to rebuild (out of module and
version tuple). Then, in parallel (or optionally sequentially), build those docker
images (lifting the existing build-docker-image job).

Finally, in swh-charts, commits a change about the new image versions to use and pushes
as a MR. That MR is to be reviewed by sysadm. They can merge it in staging and check the
behavior is appropriate. Once it is ok, they can merge it in production too. If it's not
ok, they can revert the commit.

The job can be triggered as a dry-run job so no commit nor docker images are pushed in
their respective repository or registry.

Ref. swh/infra/sysadm-environment#4724