gitlab: Ensure to rebuild open merge requests on push events
Since recently, merge requests are not rebuilt by Jenkins when pushing
new commits to it. This could be related to some default configuration
changes in latest release of GitLab plugin for Jenkins.

As that behavior is not enabled by default in gitlab trigger settings
explicitly enable it in jenkins jobs configuration.