build-docker-image: Allow to run locally through a docker agent
This introduces a new variable node which allows to either run it on the built-in
node (like actually with the production jenkins) or on the 'docker' agent node (which
should be configured either prior to choose it).

This is actually useful for running in a local jenkins and hacking around those jobs.
This will also allow further developement on similar jobs which requires to run docker.
For example, the future build-docker-images.

This also drops the instruction `updateGitlabCommitStatus` which is not needed in that
job (since that won't be opened in a mr context).

Ref. swh/infra/sysadm-environment#4724