build-docker-images: Unify behavior between built-in and local job
Prior to this commit, only the local (docker agent) mode was used. It turns
out that on built-in node, the path changes, while acceptable in local mode,
is not done the same way. This unifies such behavior by wrapping instructions
to happen in absolute path names. And now both job implementations are happy.

Ref. swh/infra/sysadm-environment#4724