incoming-tag: Add stage to trigger impacted docker images rebuild
It's installed just after the pypi upload stage.

It's not waiting for the end of this stage to finish as nothing is really
waiting for it. So debian packaging build should trigger in parallel. A quiet
period of 10s has been inserted for that new job to let pypi.org some time to
index the new version.

Refs. swh/infra/sysadm-environment#4724