build-docker-images: Make this job non concurrent
Otherwise, in some edge cases, it could happen a same dependency is needed to be built.
One of those could end up in the infinite loop of not being able to push the same
tag (since one succeeded).

It could be fixed differently but it would require a more convoluted fix. So we'll go
the simple path instead for now.

It happened recently when core and storage needed to be rebuilt and the scrubber
application got trapped in that infinite loop.