build-docker-image: Fix application tag concurrent issue push
Currently, it can still happen that concurrent build is triggered and docker images for
the same applications are built in parallel. This results in an infinite loop for one of
the docker images. [1]

This tries to overcome that issue by writing early on in the pipeline an empty tag (with
the next version to be built). That way, if concurrent issue arises, the next build
should take the next version.

[1] Possible scenario is swh.core is tagged (or any other highly connected module like
swh.storage) triggers the rebuild of the application swh.loader.git. In the mean time,
swh.loader.git is itself tagged, then possibly the 2 images are built closely enough
that one will get stuck trying to push the exact same tag.