jenkins-jobs-builder: Remove not required jenkins_token parameter
For some reason, likely due to the fact that it is a hidden one, that parameter
was not correctly transmitted to the job setting gitlab webhooks and the jenkins
jobs could not be triggered anymore.

Nevertheless, the webhooks job also has the same hidden parameter with the rotated
token as default value so we can safely remove the jenkins_token parameter for the
jenkins-jobs-builder job.