jobs/swh-docker-dev: Add trigger when pushing to swh-environment repo
Add a trigger for swh-docker-dev job when pushing to swh-environment
repository, either commit or merge request, to build swh/stack docker
image and run the docker test suites.

The docker tests will be executed only if the content of the docker
directory changed but also when job is started by a timer or manually
by a user.