stage-cypress-tests: Replace use of cypress-cloud by cypress-split
Cypress >= 13.3 blocked the use of sorry-cypress and cypress-cloud in the
cypress binary so we can no longer use these tools to execute parallel
cypress tests on Jenkins.

Fortunately, it exists another tool called cypress-split that can be used
for the same purpose without relying on a tests orchestration server.

So replace the use of sorry-cypress and cypress-cloud by cypress-split to
still benefit of parallel cypress tests in CI jobs for swh-web.

This change also brings a huge speedup for parallel cypress tests execution
as communication with a tests orchestration server is no longer required.