jobs/swh-docker-dev: Use docker agent instead of builtin node
The possibility to use docker-compose v2 has recently been added for
the docker agent so prefer to use it to run swh-environment docker
tests instead of Jenkins builtin node.

It fixes the execution of those jobs who started to fail today with
a cryptic error, bumps Python from 3.9 to 3.11 and tox from 3.x to 4.x.