Move pytest basetemp to workspace
This is a follow-up to swh/devel/swh-environment!305 which moves the
logs generated by the tests in case of errors to the temporary
directories managed by pytest.

As `archiveArtifacts` does not allow to use variables in the path
([JENKINS-5758](https://issues.jenkins.io/browse/JENKINS-5758)), we have
to change the location of the `basetemp` directory to the build
workspace instead.