docker: Fix SWH jobs execution on local Jenkins testing instance
Due to different UIDs for the jenkins user and different GIDs for
the docker group between the jenkins container and the SWH container
used to execute Jenkins pipelines, the jobs were failing due to
permission errors.

So ensure to use the same UIDs and GIDs as in SWH container for the
jenkins container and forward docker host GID in every started
container in order to fix those issues.

Also remove docker stack installation in jenkins container as it is
not needed, use docker host stack instead.