swh-pipeline: Restrict cibuildwheel builds according to tox envs
As this type of job is not dedicated to release builds, there is no
need to build all possible binary wheels when cibuildwheel is used
as we only need to check the build process does not end up with error.

Restrict the binary wheels to build with cibuildwheel to the Python
versions (being PyPy or CPython) of the tox environments defined
for the current job.