jobs/swh-pypi: Allow to build assets with missing sentry credentials
When testing jenkins jobs locally with docker, credentials used in
production are not available which makes the stage building assets
in the swh-pypi job fail due to missing sentry token.

As the sentry credentials are not required to build assets, ensure
the stage can still be executed when testing job locally.