templates/stage-python-tests: Execute each tox environment in parallel
This speedups overall pipeline execution when multiple tox environments
must be tested.

To avoid concurrency issues with tox, use a different working directory
for each environment.

The maximum number of tox environments that can be executed in parallel
can also be set through the tox_max_concurrent variable in jenkins jobs
definition. Set it to 6 by default and to 2 for jobs that already execute
tests in parallel using pytest-xdist to avoid spawning too many python
processes.