swh-pipeline: Do not use tox to build package sdist and wheel
Instead of letting tox build package sdist and wheel for each tox
environment, prefer to build the sdist and wheel once with the
build module and tell tox to install the package from it thanks
to its --installpkg option.

This should slightly speedup the tox sessions and it also allows to
remove concurrency issues when executing multiple tox environments
in parallel that were due to the tox build steps.

This also reverts 067db2cb7f4803c04030d6c0027695861a8c2630 as the
introduced hack is no longer needed.