jobs: Limit concurrent builds of jobs compiling rust code in parallel
Jobs build projects with rust code have been updated to compile it with
multiple rust versions in parallel (1.83, latest stable, nightlty).

Ensure to limit the maximum number of concurrent rust jobs to 3 in order
to avoid spawning too many rustc processes on jenkins nodes as they eat
a lot of CPU and memory.