stage-rust-*: Do not use a temporary directory as target directory
/tmp is mounted as a tmpfs in docker container running the tests but docker
agents on production jenkins have limited amount of RAM so the tmpfs is
quickly full when compiling a large rust project like swh-graph and rust
processes can also get OOM killed as a side effect.

Use a directory inside the jenkins workspace instead as before and keep
a build with the default target directory as Python tests might depend
on binaries under it.