Run cargo-clippy only once, and before cargo-build
cargo-clippy is faster, less disk-intensive, and more likely to catch a bug,
than cargo-build, so it makes sense to run it first.

Additionally, Clippy frequently introduces new checks that make the CI fail.
By only running it with Rust 1.85, we effectively pin Clippy and avoid these
unrelated failures on MRs.